Remove a GVM for deserializing nullable values
GVMs (Generic Virtual Methods) can cause large code size increases in AOT.
Because all reference types share a single code implementation, generics constrained
to class don't have this downside. This change removes a GVM for deserializing
nullable types, which includes nullable structs. Instead, deserializing nullable structs
will box.
